+++ R/loading.R
@@ -1,6 +1,6 @@
 
 #' Loading indicator to be displayed when generating screenshot
-#'
+#' 
 #' @param text Text to be displayed below loading animation.
 #' @param type Type of loading animation.
 #' @param color Color for text and loading indicator.
@@ -8,7 +8,7 @@
 #' @param size Size (in pixels).
 #' @param ... Other arguments.
 #'
-#' @return a `list` that can be used in [capture_pdf()].
+#' @return a `list` that can be used in [capture()] or [capture_pdf()].
 #' @export
 #'
 loading <- function(text = "Generating PDF, please wait...",
@@ -18,7 +18,7 @@ loading <- function(text = "Generating PDF, please wait...",
                     size = "80px",
                     ...) {
   type <- match.arg(type)
-  list(
+  dropNulls(list(
     text = text,
     type = type,
     options = list(
@@ -27,6 +27,6 @@ loading <- function(text = "Generating PDF, please wait...",
       svgColor = color,
       ...
     )
-  )
+  ))
 }