+++ examples/status-pdf.R
@@ -2,26 +2,13 @@ library(shiny)
 library(capture)
 
 ui <- fluidPage(
-  tags$h2("Capture PDF (scales) example"),
+  tags$h2("Capture PDF (loading) example"),
   capture_pdf(
     selector = "body",
-    filename = "pdf-default-scale",
-    icon("camera"), "Default scale",
-    loading = loading()
-  ),
-  capture_pdf(
-    selector = "body",
-    filename = "pdf-larger-scale",
-    icon("camera"), "Larger scale",
-    scale = 2.5,
-    loading = loading()
-  ),
-  capture_pdf(
-    selector = "body",
-    filename = "pdf-smaller-scale",
-    icon("camera"), "Smaller scale",
-    scale = 0.9,
-    loading = loading()
+    filename = "all-page",
+    icon("camera"), "Take screenshot of all page",
+    scale = 2,
+    statusInputId = "loading"
   ),
   tags$br(),
   fluidRow(
@@ -70,8 +57,16 @@ server <- function(input, output, session) {
   output$raw <- renderPrint({
     summary(distrib_r())
   })
+  
+  observeEvent(input$loading, {
+    if (identical(input$loading$status, "started")) {
+      showNotification(ui = "Capturing screenshot, please wait...")
+    }
+    if (identical(input$loading$status, "finished")) {
+      showNotification(ui = "Screenshot captured!", type = "message")
+    }
+  })
 }
 
 if (interactive())
   shinyApp(ui, server)
-